Introduce GitBasedUrlGenerator abstract class
Move the dev version detection logic to Version#isDev().
AbstractUrlGenerator class will be removed in v2.0.
