+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add support for bitbucket ssh urls ([#27](https://github.com/pyrech/composer-changelogs/pull/27))
 * Fix tests with newer composer versions ([#28](https://github.com/pyrech/composer-changelogs/pull/28))
 * Fix bug when switching from a local repository back to the original repository ([#30](https://github.com/pyrech/composer-changelogs/pull/30))
+* Add GitBasedUrlGenerator to replace AbstractUrlGenerator ([#20](https://github.com/pyrech/composer-changelogs/pull/20))
 
 ## 1.3 (2015-11-13)
